refactor(ai-engine): resolve #1746 — split logic_translator/translator.py into domain modules
Split the 983-line translator.py into focused domain mixins per Issue #1746:

- rag_context.py (RAGContextMixin): RAG pipeline integration, 55 lines
- ast_analyzer.py (ASTAnalyzerMixin): tree-sitter Java AST analysis, 179 lines
- code_translator.py (CodeTranslatorMixin): Java->JS method/class/API/event translation, 230 lines
- block_translator.py (BlockTranslatorMixin): Bedrock block/recipe JSON generation, 243 lines
- translator.py: reduced to 400-line core orchestrator that composes the mixins,
  holds shared type/enum/API mappings, singleton logic, and tool wiring

LogicTranslatorAgent now inherits all four mixins, preserving its full public
API (every method remains a bound instance method). translator.py re-exports
LogicTranslatorAgent, LLM_CODE_TEMPERATURE, TREE_SITTER_AVAILABLE, and the
mixins, so all existing imports (from agents.logic_translator.translator and
from agents.logic_translator) continue to work unchanged.

Follow-on to #1739 (tools.py split); tools.py is intentionally left untouched
as a separate wave.

Verified: 200 passed, 37 skipped on -k 'translator or logic' (identical to
baseline). No new ruff issues.
